@@ -1,8 +1,13 @@
 {
-  "name": "dogado/baum",
+  "name": "rokde/baum",
   "type": "library",
   "description": "Baum is an implementation of the Nested Set pattern for Eloquent models.",
-  "keywords": ["nested set", "laravel", "laravel 5", "eloquent", "database"],
+  "keywords": [
+    "nested set",
+    "laravel",
+    "eloquent",
+    "database"
+  ],
   "license": "MIT",
   "authors": [
     {
@@ -19,26 +24,35 @@
       "name": "Martin Seysen",
       "email": "mse@dogado.de",
       "homepage": "https://www.dogado.de"
+    },
+    {
+      "name": "Robert Kummer",
+      "email": "rok@ipunkt.biz",
+      "homepage": "https://www.ipunkt.biz"
     }
-
   ],
   "require": {
-    "php": ">=5.6.3",
-    "illuminate/console": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*",
-    "illuminate/database": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*",
-    "illuminate/events": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*",
-    "illuminate/filesystem": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*",
-    "illuminate/support": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*"
+    "php": ">=7.1",
+    "illuminate/console": "5.*|6.*",
+    "illuminate/database": "5.*|6.*",
+    "illuminate/events": "5.*|6.*",
+    "illuminate/filesystem": "5.*|6.*",
+    "illuminate/support": "5.*|6.*"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.*",
-    "mockery/mockery": "~0.9",
-    "satooshi/php-coveralls": "dev-master",
-    "fabpot/php-cs-fixer": "^1.11"
+    "phpunit/phpunit": "8.*",
+    "mockery/mockery": "~1.3"
   },
   "autoload": {
     "psr-4": {
       "Baum\\": "src/"
+    }
+  },
+  "extra": {
+    "laravel": {
+      "providers": [
+        "Baum\\Providers\\BaumServiceProvider"
+      ]
     }
   }
 }
