@@ -4,70 +4,72 @@ namespace Baum\Generators;
 
 class MigrationGenerator extends Generator
 {
-    /**
-   * Create a new migration at the given path.
-   *
-   * @param  string  $name
-   * @param  string  $path
-   *
-   * @return string
-   */
-  public function create($name, $path)
-  {
-      $path = $this->getPath($name, $path);
+	/**
+	 * Create a new migration at the given path.
+	 *
+	 * @param string $name
+	 * @param string $path
+	 *
+	 * @return string
+	 * @throws \Illuminate\Contracts\Filesystem\FileNotFoundException
+	 */
+	public function create(string $name, string $path): string
+	{
+		$path = $this->getPath($name, $path);
 
-      $stub = $this->getStub('migration');
+		$stub = $this->getStub('migration');
 
-      $this->files->put($path, $this->parseStub($stub, [
-      'table' => $this->tableize($name),
-      'class' => $this->getMigrationClassName($name),
-    ]));
+		$this->files->put($path, $this->parseStub($stub, [
+			'table' => $this->tableize($name),
+			'class' => $this->getMigrationClassName($name),
+		]));
 
-      return $path;
-  }
+		return $path;
+	}
 
-  /**
-   * Get the migration name.
-   *
-   * @param string $name
-   *
-   * @return string
-   */
-  protected function getMigrationName($name)
-  {
-      return 'create_'.$this->tableize($name).'_table';
-  }
+	/**
+	 * Get the migration name.
+	 *
+	 * @param string $name
+	 *
+	 * @return string
+	 */
+	protected function getMigrationName(string $name): string
+	{
+		return 'create_' . $this->tableize($name) . '_table';
+	}
 
-  /**
-   * Get the name for the migration class.
-   *
-   * @param string $name
-   */
-  protected function getMigrationClassName($name)
-  {
-      return $this->classify($this->getMigrationName($name));
-  }
+	/**
+	 * Get the name for the migration class.
+	 *
+	 * @param string $name
+	 * @return string
+	 */
+	protected function getMigrationClassName(string $name): string
+	{
+		return $this->classify($this->getMigrationName($name));
+	}
 
-  /**
-   * Get the full path name to the migration.
-   *
-   * @param  string  $name
-   * @param  string  $path
-   *
-   * @return string
-   */
-  protected function getPath($name, $path)
-  {
-      return $path.'/'.$this->getDatePrefix().'_'.$this->getMigrationName($name).'.php';
-  }
+	/**
+	 * Get the full path name to the migration.
+	 *
+	 * @param string $name
+	 * @param string $path
+	 *
+	 * @return string
+	 */
+	protected function getPath(string $name, string $path): string
+	{
+		return $path . DIRECTORY_SEPARATOR . $this->getDatePrefix() . '_' . $this->getMigrationName($name) . '.php';
+	}
 
-  /**
-   * Get the date prefix for the migration.
-   *
-   * @return int
-   */
-  protected function getDatePrefix()
-  {
-      return date('Y_m_d_His');
-  }
+	/**
+	 * Get the date prefix for the migration.
+	 *
+	 * @return string
+	 */
+	protected function getDatePrefix(): string
+	{
+		return date('Y_m_d_His');
+	}
 }
