@@ -4,38 +4,39 @@ namespace Baum\Generators;
 
 class ModelGenerator extends Generator
 {
-    /**
-   * Create a new model at the given path.
-   *
-   * @param  string  $name
-   * @param  string  $path
-   *
-   * @return string
-   */
-  public function create($name, $path)
-  {
-      $path = $this->getPath($name, $path);
+	/**
+	 * Create a new model at the given path.
+	 *
+	 * @param string $name
+	 * @param string $path
+	 *
+	 * @return string
+	 * @throws \Illuminate\Contracts\Filesystem\FileNotFoundException
+	 */
+	public function create(string $name, string $path): string
+	{
+		$path = $this->getPath($name, $path);
 
-      $stub = $this->getStub('model');
+		$stub = $this->getStub('model');
 
-      $this->files->put($path, $this->parseStub($stub, [
-      'table' => $this->tableize($name),
-      'class' => $this->classify($name),
-    ]));
+		$this->files->put($path, $this->parseStub($stub, [
+			'table' => $this->tableize($name),
+			'class' => $this->classify($name),
+		]));
 
-      return $path;
-  }
+		return $path;
+	}
 
-  /**
-   * Get the full path name to the migration.
-   *
-   * @param  string  $name
-   * @param  string  $path
-   *
-   * @return string
-   */
-  protected function getPath($name, $path)
-  {
-      return $path.'/'.$this->classify($name).'.php';
-  }
+	/**
+	 * Get the full path name to the migration.
+	 *
+	 * @param string $name
+	 * @param string $path
+	 *
+	 * @return string
+	 */
+	protected function getPath(string $name, string $path): string
+	{
+		return $path . DIRECTORY_SEPARATOR . $this->classify($name) . '.php';
+	}
 }
