@@ -2,7 +2,6 @@
 
 namespace Baum\Providers;
 
-use Baum\Console\BaumCommand;
 use Baum\Console\InstallCommand;
 use Baum\Generators\MigrationGenerator;
 use Baum\Generators\ModelGenerator;
@@ -10,72 +9,48 @@ use Illuminate\Support\ServiceProvider;
 
 class BaumServiceProvider extends ServiceProvider
 {
-    /**
-     * Baum version.
-     *
-     * @var string
-     */
-    const VERSION = '1.3.0';
+	/**
+	 * Baum version.
+	 *
+	 * @var string
+	 */
+	const VERSION = '2.0.0';
 
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        $this->registerCommands();
-    }
+	public function register()
+	{
+		if (!$this->app->runningInConsole()) {
+			return;
+		}
 
-    /**
-     * Register the commands.
-     *
-     * @return void
-     */
-    public function registerCommands()
-    {
-        $this->registerBaumCommand();
-        $this->registerInstallCommand();
+		$this->registerInstallCommand();
 
-        // Resolve the commands with Artisan by attaching the event listener to Artisan's
-        // startup. This allows us to use the commands from our terminal.
-        $this->commands('command.baum', 'command.baum.install');
-    }
+		// Resolve the commands with Artisan by attaching the event listener to Artisan's
+		// startup. This allows us to use the commands from our terminal.
+		$this->commands('command.baum.install');
+	}
 
-    /**
-     * Register the 'baum' command.
-     *
-     * @return void
-     */
-    protected function registerBaumCommand()
-    {
-        $this->app->singleton('command.baum', function ($app) {
-            return new BaumCommand();
-        });
-    }
+	/**
+	 * Register the 'baum:install' command.
+	 *
+	 * @return void
+	 */
+	protected function registerInstallCommand()
+	{
+		$this->app->singleton('command.baum.install', function ($app) {
+			$migrator = new MigrationGenerator($app['files']);
+			$modeler = new ModelGenerator($app['files']);
 
-    /**
-     * Register the 'baum:install' command.
-     *
-     * @return void
-     */
-    protected function registerInstallCommand()
-    {
-        $this->app->singleton('command.baum.install', function ($app) {
-            $migrator = new MigrationGenerator($app['files']);
-            $modeler = new ModelGenerator($app['files']);
+			return new InstallCommand($migrator, $modeler);
+		});
+	}
 
-            return new InstallCommand($migrator, $modeler);
-        });
-    }
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return ['command.baum', 'command.baum.install'];
-    }
+	/**
+	 * Get the services provided by the provider.
+	 *
+	 * @return array
+	 */
+	public function provides()
+	{
+		return ['command.baum.install'];
+	}
 }
