@@ -4,69 +4,69 @@ namespace Baum;
 
 class SetValidator
 {
-    /**
-     * Node instance for reference.
-     *
-     * @var Node
-     */
-    protected $node = null;
+	/**
+	 * Node instance for reference.
+	 *
+	 * @var Node
+	 */
+	protected $node = null;
 
-    /**
-     * Create a new \Baum\SetValidator class instance.
-     *
-     * @param Node $node
-     */
-    public function __construct(Node $node)
-    {
-        $this->node = $node;
-    }
+	/**
+	 * Create a new \Baum\SetValidator class instance.
+	 *
+	 * @param Node $node
+	 */
+	public function __construct(Node $node)
+	{
+		$this->node = $node;
+	}
 
-    /**
-     * Determine if the validation passes.
-     *
-     * @return bool
-     */
-    public function passes()
-    {
-        return $this->validateBounds() &&
-            $this->validateDuplicates() &&
-            $this->validateRoots();
-    }
+	/**
+	 * Determine if the validation passes.
+	 *
+	 * @return bool
+	 */
+	public function passes(): bool
+	{
+		return $this->validateBounds() &&
+			$this->validateDuplicates() &&
+			$this->validateRoots();
+	}
 
-    /**
-     * Determine if validation fails.
-     *
-     * @return bool
-     */
-    public function fails()
-    {
-        return !$this->passes();
-    }
+	/**
+	 * Determine if validation fails.
+	 *
+	 * @return bool
+	 */
+	public function fails(): bool
+	{
+		return !$this->passes();
+	}
 
-    /**
-     * Validates bounds of the nested tree structure. It will perform checks on
-     * the `lft`, `rgt` and `parent_id` columns. Mainly that they're not null,
-     * rights greater than lefts, and that they're within the bounds of the parent.
-     *
-     * @return bool
-     */
-    protected function validateBounds()
-    {
-        $connection = $this->node->getConnection();
-        $grammar = $connection->getQueryGrammar();
+	/**
+	 * Validates bounds of the nested tree structure. It will perform checks on
+	 * the `lft`, `rgt` and `parent_id` columns. Mainly that they're not null,
+	 * rights greater than lefts, and that they're within the bounds of the parent.
+	 *
+	 * @return bool
+	 */
+	protected function validateBounds(): bool
+	{
+		$connection = $this->node->getConnection();
+		$grammar = $connection->getQueryGrammar();
 
-        $tableName = $this->node->getTable();
-        $primaryKeyName = $this->node->getKeyName();
-        $parentColumn = $this->node->getQualifiedParentColumnName();
+		$tableName = $this->node->getTable();
+		$primaryKeyName = $this->node->getKeyName();
+		$parentColumn = $this->node->getQualifiedParentColumnName();
 
-        $lftCol = $grammar->wrap($this->node->getLeftColumnName());
-        $rgtCol = $grammar->wrap($this->node->getRightColumnName());
+		$lftCol = $grammar->wrap($this->node->getLeftColumnName());
+		$rgtCol = $grammar->wrap($this->node->getRightColumnName());
 
-        $qualifiedLftCol = $grammar->wrap($this->node->getQualifiedLeftColumnName());
-        $qualifiedRgtCol = $grammar->wrap($this->node->getQualifiedRightColumnName());
-        $qualifiedParentCol = $grammar->wrap($this->node->getQualifiedParentColumnName());
+		$qualifiedLftCol = $grammar->wrap($this->node->getQualifiedLeftColumnName());
+		$qualifiedRgtCol = $grammar->wrap($this->node->getQualifiedRightColumnName());
+		$qualifiedParentCol = $grammar->wrap($this->node->getQualifiedParentColumnName());
 
-        $whereStm = "(
+		$whereStm = "(
             $qualifiedLftCol IS NULL OR
             $qualifiedRgtCol IS NULL OR
             $qualifiedLftCol >= $qualifiedRgtCol OR
@@ -79,172 +79,172 @@ class SetValidator
             )
         )";
 
-        $query = $this->node->newQuery()
-            ->join($connection->raw($grammar->wrapTable($tableName) . ' AS parent'),
-                $parentColumn, '=', $connection->raw('parent.' . $grammar->wrap($primaryKeyName)),
-                'left outer')
-            ->whereRaw($whereStm);
+		$query = $this->node->newQuery()
+			->join($connection->raw($grammar->wrapTable($tableName) . ' AS parent'),
+				$parentColumn, '=', $connection->raw('parent.' . $grammar->wrap($primaryKeyName)),
+				'left outer')
+			->whereRaw($whereStm);
 
-        return $query->count() == 0;
-    }
+		return $query->count() == 0;
+	}
 
-    /**
-     * Checks that there are no duplicates for the `lft` and `rgt` columns.
-     *
-     * @return bool
-     */
-    protected function validateDuplicates()
-    {
-        return
-            !$this->duplicatesExistForColumn($this->node->getQualifiedLeftColumnName()) &&
-            !$this->duplicatesExistForColumn($this->node->getQualifiedRightColumnName());
-    }
+	/**
+	 * Checks that there are no duplicates for the `lft` and `rgt` columns.
+	 *
+	 * @return bool
+	 */
+	protected function validateDuplicates(): bool
+	{
+		return
+			!$this->duplicatesExistForColumn($this->node->getQualifiedLeftColumnName()) &&
+			!$this->duplicatesExistForColumn($this->node->getQualifiedRightColumnName());
+	}
 
-    /**
-     * For each root of the whole nested set tree structure, checks that their
-     * `lft` and `rgt` bounds are properly set.
-     *
-     * @return bool
-     */
-    protected function validateRoots()
-    {
-        $roots = forward_static_call([get_class($this->node), 'roots'])->get();
+	/**
+	 * For each root of the whole nested set tree structure, checks that their
+	 * `lft` and `rgt` bounds are properly set.
+	 *
+	 * @return bool
+	 */
+	protected function validateRoots(): bool
+	{
+		$roots = forward_static_call([get_class($this->node), 'roots'])->get();
 
-        // If a scope is defined in the model we should check that the roots are
-        // valid *for each* value in the scope columns.
-        if ($this->node->isScoped()) {
-            return $this->validateRootsByScope($roots);
-        }
+		// If a scope is defined in the model we should check that the roots are
+		// valid *for each* value in the scope columns.
+		if ($this->node->isScoped()) {
+			return $this->validateRootsByScope($roots);
+		}
 
-        return $this->isEachRootValid($roots);
-    }
+		return $this->isEachRootValid($roots);
+	}
 
-    /**
-     * Checks if duplicate values for the column specified exist. Takes
-     * the Nested Set scope columns into account (if appropiate).
-     *
-     * @param   string $column
-     *
-     * @return  bool
-     */
-    protected function duplicatesExistForColumn($column)
-    {
-        $connection = $this->node->getConnection();
-        $grammar = $connection->getQueryGrammar();
+	/**
+	 * Checks if duplicate values for the column specified exist. Takes
+	 * the Nested Set scope columns into account (if appropiate).
+	 *
+	 * @param string $column
+	 *
+	 * @return  bool
+	 */
+	protected function duplicatesExistForColumn($column): bool
+	{
+		$connection = $this->node->getConnection();
+		$grammar = $connection->getQueryGrammar();
 
-        $columns = array_merge($this->node->getQualifiedScopedColumns(), [$column]);
+		$columns = array_merge($this->node->getQualifiedScopedColumns(), [$column]);
 
-        $columnsForSelect = implode(', ', array_map(function ($col) use ($grammar) {
-            return $grammar->wrap($col);
-        }, $columns));
+		$columnsForSelect = implode(', ', array_map(function ($col) use ($grammar) {
+			return $grammar->wrap($col);
+		}, $columns));
 
-        $wrappedColumn = $grammar->wrap($column);
+		$wrappedColumn = $grammar->wrap($column);
 
-        $query = $this->node->newQuery()
-            ->select($connection->raw("$columnsForSelect, COUNT($wrappedColumn)"))
-            ->havingRaw("COUNT($wrappedColumn) > 1");
+		$query = $this->node->newQuery()
+			->select($connection->raw("$columnsForSelect, COUNT($wrappedColumn)"))
+			->havingRaw("COUNT($wrappedColumn) > 1");
 
-        foreach ($columns as $col) {
-            $query->groupBy($col);
-        }
+		foreach ($columns as $col) {
+			$query->groupBy($col);
+		}
 
-        $result = $query->first();
+		$result = $query->first();
 
-        return !is_null($result);
-    }
+		return !is_null($result);
+	}
 
-    /**
-     * Check that each root node in the list supplied satisfies that its bounds
-     * values (lft, rgt indexes) are less than the next.
-     *
-     * @param   mixed $roots
-     *
-     * @return  bool
-     */
-    protected function isEachRootValid($roots)
-    {
-        $left = $right = 0;
+	/**
+	 * Check that each root node in the list supplied satisfies that its bounds
+	 * values (lft, rgt indexes) are less than the next.
+	 *
+	 * @param mixed $roots
+	 *
+	 * @return  bool
+	 */
+	protected function isEachRootValid($roots): bool
+	{
+		$left = $right = 0;
 
-        foreach ($roots as $root) {
-            $rootLeft = $root->getLeft();
-            $rootRight = $root->getRight();
+		foreach ($roots as $root) {
+			$rootLeft = $root->getLeft();
+			$rootRight = $root->getRight();
 
-            if (!($rootLeft > $left && $rootRight > $right)) {
-                return false;
-            }
+			if (!($rootLeft > $left && $rootRight > $right)) {
+				return false;
+			}
 
-            $left = $rootLeft;
-            $right = $rootRight;
-        }
+			$left = $rootLeft;
+			$right = $rootRight;
+		}
 
-        return true;
-    }
+		return true;
+	}
 
-    /**
-     * Check that each root node in the list supplied satisfies that its bounds
-     * values (lft, rgt indexes) are less than the next *within each scope*.
-     *
-     * @param   mixed $roots
-     *
-     * @return  bool
-     */
-    protected function validateRootsByScope($roots)
-    {
-        foreach ($this->groupRootsByScope($roots) as $scope => $groupedRoots) {
-            $valid = $this->isEachRootValid($groupedRoots);
+	/**
+	 * Check that each root node in the list supplied satisfies that its bounds
+	 * values (lft, rgt indexes) are less than the next *within each scope*.
+	 *
+	 * @param mixed $roots
+	 *
+	 * @return  bool
+	 */
+	protected function validateRootsByScope($roots): bool
+	{
+		foreach ($this->groupRootsByScope($roots) as $scope => $groupedRoots) {
+			$valid = $this->isEachRootValid($groupedRoots);
 
-            if (!$valid) {
-                return false;
-            }
-        }
+			if (!$valid) {
+				return false;
+			}
+		}
 
-        return true;
-    }
+		return true;
+	}
 
-    /**
-     * Given a list of root nodes, it returns an array in which the keys are the
-     * array of the actual scope column values and the values are the root nodes
-     * inside that scope themselves.
-     *
-     * @param   mixed $roots
-     *
-     * @return  array
-     */
-    protected function groupRootsByScope($roots)
-    {
-        $rootsGroupedByScope = [];
+	/**
+	 * Given a list of root nodes, it returns an array in which the keys are the
+	 * array of the actual scope column values and the values are the root nodes
+	 * inside that scope themselves.
+	 *
+	 * @param mixed $roots
+	 *
+	 * @return  array
+	 */
+	protected function groupRootsByScope($roots): array
+	{
+		$rootsGroupedByScope = [];
 
-        foreach ($roots as $root) {
-            $key = $this->keyForScope($root);
+		foreach ($roots as $root) {
+			$key = $this->keyForScope($root);
 
-            if (!isset($rootsGroupedByScope[$key])) {
-                $rootsGroupedByScope[$key] = [];
-            }
+			if (!isset($rootsGroupedByScope[$key])) {
+				$rootsGroupedByScope[$key] = [];
+			}
 
-            $rootsGroupedByScope[$key][] = $root;
-        }
+			$rootsGroupedByScope[$key][] = $root;
+		}
 
-        return $rootsGroupedByScope;
-    }
+		return $rootsGroupedByScope;
+	}
 
-    /**
-     * Builds a single string for the given scope columns values. Useful for
-     * making array keys for grouping.
-     *
-     * @param Node $node
-     *
-     * @return string
-     */
-    protected function keyForScope($node)
-    {
-        return implode('-', array_map(function ($column) use ($node) {
-            $value = $node->getAttribute($column);
+	/**
+	 * Builds a single string for the given scope columns values. Useful for
+	 * making array keys for grouping.
+	 *
+	 * @param Node $node
+	 *
+	 * @return string
+	 */
+	protected function keyForScope(Node $node): string
+	{
+		return implode('-', array_map(function ($column) use ($node) {
+			$value = $node->getAttribute($column);
 
-            if (is_null($value)) {
-                return 'NULL';
-            }
+			if (is_null($value)) {
+				return 'NULL';
+			}
 
-            return $value;
-        }, $node->getScopedColumns()));
-    }
+			return $value;
+		}, $node->getScopedColumns()));
+	}
 }
