@@ -1,9 +1,7 @@
 <?php
 
-$config = [
-    'driver'    => 'sqlite',
-    'database'  => ':memory:',
-    'prefix'    => '',
+return [
+	'driver' => 'sqlite',
+	'database' => ':memory:',
+	'prefix' => '',
 ];
-
-return $config;
