@@ -5,38 +5,38 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Category extends Node
 {
-    protected $table = 'categories';
+	protected $table = 'categories';
 
-    public $timestamps = false;
+	public $timestamps = false;
 }
 
 class ScopedCategory extends Category
 {
-    protected $scoped = ['company_id'];
+	protected $scoped = ['company_id'];
 }
 
 class MultiScopedCategory extends Category
 {
-    protected $scoped = ['company_id', 'language'];
+	protected $scoped = ['company_id', 'language'];
 }
 
 class OrderedCategory extends Category
 {
-    protected $orderColumn = 'name';
+	protected $orderColumn = 'name';
 }
 
 class OrderedScopedCategory extends Category
 {
-    protected $scoped = ['company_id'];
+	protected $scoped = ['company_id'];
 
-    protected $orderColumn = 'name';
+	protected $orderColumn = 'name';
 }
 
 class SoftCategory extends Category
 {
-    use SoftDeletes;
+	use SoftDeletes;
 
-    public $timestamps = true;
+	public $timestamps = true;
 
-    protected $dates = ['deleted_at'];
+	protected $dates = ['deleted_at'];
 }
