@@ -4,54 +4,56 @@ use Mockery as m;
 
 class CategoryCustomEventsTest extends CategoryTestCase
 {
-    public function tearDown()
-    {
-        m::close();
-    }
+	public function tearDown(): void
+	{
+		m::close();
+	}
 
-    public function testMovementEventsFire()
-    {
-        $dispatcher = Category::getEventDispatcher();
-        Category::setEventDispatcher($events = m::mock('Illuminate\Events\Dispatcher'));
+	public function testMovementEventsFire()
+	{
+		$dispatcher = Category::getEventDispatcher();
+		Category::setEventDispatcher($events = m::mock('Illuminate\Events\Dispatcher'));
+		$events->shouldReceive('dispatch')->andReturns();
 
-        $child = $this->categories('Child 1');
+		/** @var \Category $child */
+		$child = $this->categories('Child 1');
 
-        $events->shouldReceive('until')->once()->with('eloquent.moving: '.get_class($child), $child)->andReturn(true);
-        $events->shouldReceive('fire')->once()->with('eloquent.moved: '.get_class($child), $child)->andReturn(true);
+		$events->shouldReceive('until')->once()->with('eloquent.moving: ' . get_class($child), $child)->andReturn(true);
 
-        $child->moveToRightOf($this->categories('Child 3'));
+		$child->moveToRightOf($this->categories('Child 3'));
 
-        Category::unsetEventDispatcher();
-        Category::setEventDispatcher($dispatcher);
-    }
+		Category::unsetEventDispatcher();
+		Category::setEventDispatcher($dispatcher);
+	}
 
-    public function testMovementHaltsWhenReturningFalseFromMoving()
-    {
-        $unchanged = $this->categories('Child 2');
+	public function testMovementHaltsWhenReturningFalseFromMoving()
+	{
+		$unchanged = $this->categories('Child 2');
 
-        $dispatcher = Category::getEventDispatcher();
+		$dispatcher = Category::getEventDispatcher();
 
-        Category::setEventDispatcher($events = m::mock('Illuminate\Events\Dispatcher[until]'));
-        $events->shouldReceive('until')->once()->with('eloquent.moving: '.get_class($unchanged), $unchanged)->andReturn(false);
+		Category::setEventDispatcher($events = m::mock('Illuminate\Events\Dispatcher[until]'));
+		$events->shouldReceive('until')->once()->with('eloquent.moving: ' . get_class($unchanged),
+			$unchanged)->andReturn(false);
 
-    // Force "moving" to return false
-    Category::moving(function ($node) {
-        return false;
-    });
+		// Force "moving" to return false
+		Category::moving(function ($node) {
+			return false;
+		});
 
-        $unchanged->makeRoot();
+		$unchanged->makeRoot();
 
-        $unchanged->reload();
+		$unchanged->reload();
 
-        $this->assertEquals(1, $unchanged->getParentId());
-        $this->assertEquals(1, $unchanged->getLevel());
-        $this->assertEquals(4, $unchanged->getLeft());
-        $this->assertEquals(7, $unchanged->getRight());
+		$this->assertEquals(1, $unchanged->getParentId());
+		$this->assertEquals(1, $unchanged->getLevel());
+		$this->assertEquals(4, $unchanged->getLeft());
+		$this->assertEquals(7, $unchanged->getRight());
 
-    // Restore
-    Category::getEventDispatcher()->forget('eloquent.moving: '.get_class($unchanged));
+		// Restore
+		Category::getEventDispatcher()->forget('eloquent.moving: ' . get_class($unchanged));
 
-        Category::unsetEventDispatcher();
-        Category::setEventDispatcher($dispatcher);
-    }
+		Category::unsetEventDispatcher();
+		Category::setEventDispatcher($dispatcher);
+	}
 }
