@@ -2,18 +2,23 @@
 
 class CategoryTestCase extends BaumTestCase
 {
-    public static function setUpBeforeClass()
-    {
-        with(new CategoryMigrator())->up();
-    }
+	public static function setUpBeforeClass(): void
+	{
+		with(new CategoryMigrator())->up();
+	}
 
-    public function setUp()
-    {
-        with(new CategorySeeder())->run();
-    }
+	public function setUp(): void
+	{
+		with(new CategorySeeder())->run();
+	}
 
-    protected function categories($name, $className = 'Category')
-    {
-        return forward_static_call_array([$className, 'where'], ['name', '=', $name])->first();
-    }
+	/**
+	 * @param $name
+	 * @param string $className
+	 * @return \Category|\Baum\Node[]
+	 */
+	protected function categories($name, $className = 'Category')
+	{
+		return forward_static_call_array([$className, 'where'], ['name', '=', $name])->first();
+	}
 }
