@@ -2,12 +2,12 @@
 
 class ClusterTestCase extends BaumTestCase
 {
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         with(new ClusterMigrator())->up();
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         with(new ClusterSeeder())->run();
     }
